cryptodev: do not cache file descriptor in 'open'
The file descriptor returned by get_dev_crypto is cached after a
successful return. The issue is, it is cached inside 'open_dev_crypto'
which is no longer useful as a general purpose open("/dev/crypto")
function.

This patch is a refactoring that moves the caching operation from
open_dev_crypto to get_dev_crypto and leaves the former as a simpler
function true to its name

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>